Move app to separate file
